gen: implement binary arithmetic
This also adds an "indirect" flag to qbe_value, which is used to
indicate that a value stores its result "indirectly". This is used by
gen_store to choose between copy and store instructions.

I suspected this would be necessary at some point, but I'm not yet sure
how exactly it's going to hold up as gen expands. Keeping an eye on it.