Implement aggregate type assignments
This is still a bit messy and I think there's room for simplificaiton. I
choose to copy via qbe instructions, which gives qbe an opportunity to
optimize it out if we're not using it - something which the memcpy
approach would not have been able to do.

One change this makes is to function call parameters - and one possible
improvement would be to add some kind of "alloc if necssary" flag to
qbe_value, which gen_store would read and allocate the appropriate
amount of space if necessary before copying the data in. This would
remove one useless stack allocation when not passing array literals
directly into function parameters. However, I suspect that a simpler
approach to aggregate types in general is possible, and would save us
the trouble.