gen: introduce _with gen_expr construct
This allows us to merge the non-at and -at implementations of
gen_expr_list by introducing a new -with variant, for functions which
don't have a preferred allocation paradigm.

This gives us the best of both worlds, such that:

	let x = {
		10;
	};
	let y = {
		coords { ... };
	};
	{
		coords { ... };
	};

All use the optimal allocation strategy. And so does this:

	foobar(match (x) {
		a => y,
		b => coords { ... },
	});

Only one object is allocated and there are zero copies.

In summary:

- gen_expr_* variants use callee-allocated storage and return their
  result to the caller.
- gen_expr_*_at variants use caller-allocated storage, and write their
  result into that storage.
- gen_expr_*_with variants do not have a preference, and will use
  caller-allocated storage if available, or return the result if not.

gen_expr() shims -at only functions by allocating storage for them.
gen_expr_at() shims non-at functions by writing their result to the
caller-provided storage. Both functions use -with functions in similar
ways, either providing or not providing a place to store the result.

Signed-off-by: Drew DeVault <sir@cmpwn.com>