gen: use 'from' for inner union in cast-to-tagged
In this situation, 'from' is always a more specific union than 'to',
which means that it will always have a size equal to or less than 'to'.
If we based the inner type on the 'to' type, we could end up with a
larger type, which would cause out-of-bounds reads from the source
object.

The new approach might have issues on big-endian systems in the future;
we'll revisit it if we ever add a BE target.

Signed-off-by: Drew DeVault <sir@cmpwn.com>