gen: fix nested match with aliases
In this case:

type t = (void | int);
fn foo() void = {
	let x: (t | int) = 10;
	match (x) {
		(void | int) => void,
		int => void,
	};
};

we wouldn't exit the loop on test == t because it doesn't have the same
id as (void | int). Rather than just testing against t, test against
type_dealias(t) as well.

We don't want to dealias _case->type, because that can lead to false
positives. For example, in this case:

type foo = void;
type bar = void;
fn foo() void = {
	match (foo) {
		bar => void,
		foo => void,
	};
};

we need to avoid matching with bar.

This issue was causing 18-match to fail, now that it's fixed, enable
that test.

Signed-off-by: Eyal Sawady <ecs@d2evs.net>