gen: fix out-of-bounds write in cast to narrower union for match
If we have something like

	export fn f() (void | i32) = {
	        let x: (void | i32 | i64) = 1234i32;
	        return match (x) {
	                y: (void | i32) => y,
	                * => abort(),
	        };
	};

the types in the inner union of the `to` type are a subset of the
types in the inner union of the `from` type. However, harec was
generating a copy of the full `from` union, which results in writing
past the end of the storage allocated for the `to` union, as well
as unaligned stores.

To fix this, use the narrower union type for the generated copy.

Fixes tests/18-match on riscv64.