type_store: don't initialize types that can't be referenced
All derivative types of a type alias will always have set a superset of
flags the alias was defined with.

Consider `type foo = !void;`.
Such a definition leaves no way to create a value of type "foo, but
without the TYPE_ERROR type flag", so there is no need to put such a
type into the type store. Similar logic applies to const flag, so we can
just ignore any flag combination that is a strict subset of the flags of
the alias-type's secondary type.

Signed-off-by: Bor Grošelj Simić <bor.groseljsimic@telemach.net>