Overhaul allocation expressions
Prior to this change, allocation expressions changed their behavior
dramatically depending on context. This change significantly reduces the
effect of the type hint on the behavior of alloc, and uses distinct
grammatical constructs to address each case instead.

Allocations are now broken down into four categories:

- Objects		// alloc(42)
- Slice with capacity	// alloc([], 42);
- Slice with length	// alloc([0...], 42);
- Slice copies		// alloc(x...)

All of these previously lived under the same code path and used mostly
the same grammar, and had a lot of hacks to distinguish them. Now they
are distinguished during parse, except cap/length, which is done in
check, and is made easier/less hacky thanks to the previous array
expandability change.

There is likely to be at least one additional change to allocations in
the future: presently the nullable flag in the type hint is used to
determine if the allocation should abort on allocation failure or not.
In the future, we are likely to make OOM handling mandatory, so this
will change.

I also think there's a potential footgun in the alloc(x...) case,
because in some cases `alloc(x)` will compile but do something
unexpected -- allocate a *[]int which refers to the same underlying data
source, instead of creating a []int which allocates a separate copy of
the underlying data. It may be wise to revisit this later.

Signed-off-by: Drew DeVault <sir@cmpwn.com>