Overhaul expandable arrays
This improves the design of expandable arrays somewhat so that we are
able to use them in a broader set of contexts, notably for the coming
allocation improvements.

Essentially, this adds the "expandable" tag to the array type itself, so
that an expression of [1, 2, 3...] becomes a [3]int with the expandable
tag set, rather than the previous behavior of looking into the type hint
to find a suitable length.

This is an ABI breaking change.

Signed-off-by: Drew DeVault <sir@cmpwn.com>