check: ensure constant slice-alloc capacity >= length
For slice allocation, checks at compile time that the capacity is
greater than or equal to the length of the initializer, if both the
capacity and the initializer are part of the translation-compatible
subset. No runtime checks are included in this commit.

References: https://todo.sr.ht/~sircmpwn/hare/477
References: https://todo.sr.ht/~sircmpwn/hare/581
Signed-off-by: Sebastian <sebastian@sebsite.pw>