make enum type id algorithm rely on enum name
Before this change enum type id was calculated based on names and values
of enum field. The enum type wasn't really useful since the only way to
access it's values was through a type alias. The old method also
rendered some complex combinations of enum declarations and constant
declarations impossible to resolve.

This is now solved by special-casing enum types in the grammar. Unlike
other types they are not standalone and can only be used in type
declarations. The name given to an enum in an enum declaration is not
just an alias, but is a property of the enum type. The type id algorithm
for enums now relies on that name.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>