simplify and generalize declaration resolution
Previous version of declaration resolution was algorithmically correct,
but implemented in a pretty convoluted way, which made it difficult to
understand and reason about. It mainly addressed the needs of
(co)recursive type resolution and all other functionality was sort of
glued onto it with limited success.

This patch makes "scan_decl_finish" into merely a wrapper that takes
care of context setup and teardown and detects circular dependencies,
the main resolution functionality is left to arbitrary resolver
functions that can be passed to it.

This will among other things allow us to use the mechanism for enum
alias resolution which is a separate procedure, but needs to take care
of context setup and circular dependencies in the same way.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>