gen: fix segfault when allocating multiple of 2^32
In QBE, the jnz instruction only checks the lower 32 bits of its first
argument (i.e. it's treated as a word). Since the passed size was being
checked directly, any multiple of 2^32 would be treated as zero,
resulting in a segfault at runtime.

	let n: size = 64 * 64 * 1024 * 1024;
	let x: []u8 = alloc([0...], n);

This has been fixed by first using cnel to compare the size to zero, and
then using the result of that in the jnz instruction.

Fixes: https://todo.sr.ht/~sircmpwn/hare/739
Signed-off-by: Sebastian <sebastian@sebsite.pw>