gen: don't directly use jnz on longs
jnz always operates on words. Because of this, there are rare cases
where the lower 32 bits of an integer are zero, but the upper 32 bits
aren't, so the code branches incorrectly. This is fixed by using cnel
(or a similar comparison function) first, and then using the return
value of that as the first operand to jnz.

Signed-off-by: Sebastian <sebastian@sebsite.pw>