Allow abort/assert to accept non-constant strings.
This is useful for tests, allowing for statements like:
    assert(a == b, fmt::asprintf("expected {}, got {}, a, b));

The message for the above assertion (if a = 4, b = 5) will be:
    Abort: /path/to/file.ha:42:1: expected 4, got 5

This patch reorders the output for assert(false), without the
message string, to be consistent with assert(false, "foo").
Previously it was "Abort: Assertion failed: location", but now it
is "Abort: location: Assertion failed" to match "Abort: location:
foo".

This patch also changes the signature for rt::_abort, adding a
location parameter (matching abort_fixed).

Related to ~sircmpwn/hare#109.

Signed-off-by: Joe Finney <me@spxtr.net>