check: more strict rules for type asserting nullable pointers
The compiler used to not check what the target type of the type
assertion of a nullable pointer was. Because of this, given x is a
nullable *int, expressions like `x as *size` and `x as nullable *int`
were incorrectly allowed. Furthermore, expressions like `x as int` and
`x as size` resulted in an abort() in gen.c.

This has been corrected in accordance with the spec, so now nullable
pointers may be type asserted to null or to a non-nullable pointer type
with the same referent type (i.e. x can now only be type asserted to
null and *int).

Signed-off-by: Sebastian <sebastian@sebsite.pw>