check: fix type reduction when yielding without value
Previously, using `yield` without a value didn't correctly add a `void`
expression to the list of yields for a scope. So, for instance, code
like this would incorrectly compile:

	let x: int = {
		if (true) yield 0;
		yield;
	};

If the above `yield` was changed to a `yield void` this would correctly
not compile. This has been fixed, by adding a `void` value to a yield
without a value during check.

Signed-off-by: Sebastian <sebastian@sebsite.pw>