check: fix NULL-dereference bugs
type_dereference returns NULL if the pointer is nullable. Some parts of
check didn't properly test for this, resulting in a segfault when a
nullable pointer was used. In particular, append/insert and delete
expressions. Furthermore, type_dereference was used on the type hint of
an object allocation expression, which was unnecessary and was also
susceptible to the segfault.

Signed-off-by: Sebastian <sebastian@sebsite.pw>