promote_const: fix ambiguous tagged unions
Specifically fix a crash on ambiguous tagged unions nested inside an
otherwise-unambiguous union where the ambiguous union has the highest
type ID of the members of the unambiguous union and the default type for
the constant isn't assignable to the unambiguous union. In this case,
the recursive promote_const goes through the lower_const(b, NULL)
codepath, so we need to make sure to re-lower b to the tag.

If the tag is NULL, this leads to b incorrectly being lowered to its
default type. However, in this case, either b will get re-lowered and
the temporary lowering to the default type doesn't matter or it won't
and promote_const will return NULL anyways, so we'll error out before
that matters.

Signed-off-by: Ember Sawady <ecs@d2evs.net>