rewrite lex_literal
The old version grew another boolean variable for the state with every new
feature that was added. The state is now condensed into base information
plus 4 flags that work together nicely. Actual conversion to numbers is
also greatly simplified and the memory issues that the old one had are
avoided.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>