check: handle errors without (ab)using type hints
Currently, errors are mandatory when the expression is given a void type
hint. This is a hack, and will need to be changed to allow for zero-size
bindings. The spec just, doesn't say anything about this, this behavior
just isn't specified, so I had to decide what the Correct behavior is
here. I decided that errors must be handled when the result of the
expression is discarded, which only happens within compound expressions,
for expressions, and defer expressions.

Signed-off-by: Sebastian <sebastian@sebsite.pw>