check: add implicit yield at end of compound expr
This is necessary when the expression is yielded from before the end,
but not at the end. For instance:

	let x: (int | void) = {
		if (false) yield 42;
	};

Previously, there were multiple issues here. As written, this would
abort in qtype_lookup. Furthermore, if the type hint was removed (and 42
was given an explicit suffix like i32), this would miscompile, with x
being only an i32, and equaling 42 despite that branch being dead.

Fixes: https://todo.sr.ht/~sircmpwn/hare/785
Signed-off-by: Sebastian <sebastian@sebsite.pw>