Set slice capacity to (source length) - L.
The spec says that the capacity of a slice s[L..H] should be the length
of the source object minus L. If the length of the source is undefined,
the capacity should be H - L. At some point in the last year, this
broke for slices borrowing from arrays, which now took capacity H - L
even if the starting array had some defined length.

This would break bounds checks in static append to slices of 0 length
but (what should have been) nonzero capacity:
    let buf: [5]int = [0...];
    let slice = buf[2..2]; // Should have length 0, capacity 3.
    static append(slice, 1); // Should be okay, but wasn't.
among other things.

This patch hopefully fixes it!

Signed-off-by: Joe Finney <me@spxtr.net>