type_store: sort struct fields by @offset
So, ok: struct fields need to be sorted by offset, so they're emitted
correctly in the IR and in typedefs. Currently, struct fields aren't
sorted this way when the list is first constructed; they're sorted
within qtype.c and typedef.c (separately). Now that we have the
invariant that struct fields in source code must have strictly
increasing offsets, we can get rid of the sorting entirely, and just
construct the list in order of offsets, so it's already sorted. The main
motivation here is code cleanliness.

Except, there's one slight problem here: currently, fields are sorted
alphabetically at first. This is used when constructing the list, to
check for duplicate field names. Sorting by offset requires that the
entire list thus far is checked for every additional field added.

However, this isn't actually an issue in practice. The list was already
being walked sequentially, since it's a linked list, so a binary search
isn't really possible. So the change here just makes it so the entire
list has to be walked, rather than stopping when a field is
alphabetically greater than the one currently being checked. Also worth
noting that there's already an O(n^2) algorithm being used when checking
if a struct initializer is exhaustive.

In practice, this doesn't have any noticeable performance implications.
Even a struct with 1000 fields is still checked near-instantly (in human
timescales, that is). A struct with 10000 fields takes significantly
longer; around a second or so, but this should be fixed by not using
linked lists here, since they're really a bad data structure to use here
(or anywhere in harec, really).

Signed-off-by: Sebastian <sebastian@sebsite.pw>