genutil: remove xstrdup from mkrtfunc
rtfuncs are stored separately from other qbe_values, so we'll still be
able to free the others without this causing any issues.

While here also remove it from the header and make it static, since it's
only ever used in rtfunc_init. This used to be used throughout gen.c,
but it no longer is.

Signed-off-by: Sebastian <sebastian@sebsite.pw>