parse, check: update imports
The following import forms are removed:

	use foo = bar::{baz}; // alias + members
	use foo::{bar = baz}; // aliased members

The alias form's syntax is also changed to match the spec: the alias now
must be a name rather than an identifier.

The logic for importing enum members in non-members forms is also
changed: the previous implementation checked if the scope object was a
constant with an enum type, and, if it was, inserted the last component
of the identifier namespace into the newly imported identifier. That is,
the object foo::bar::baz imported as bar::baz, but foo::bar::baz::BAT
imported as bar::baz::BAT if BAT had an enum type. The problem is that
this incorrectly also happened for non-enum-constants with an enum type.
That is, foo::BAR would import as foo::foo::BAR if BAR was defined with
an enum type. This has been fixed by checking if the identifier's
namespace themselves identify an object in the module's scope, and
requiring that this object be an enum type.

The normal import form still also imports fully-qualified identifiers.
This behavior is currently depended upon by typedefs.

References: https://lists.sr.ht/~sircmpwn/hare-rfc/<CW9K5L4ALVKU.1Y541JKBOIPRU%40notmylaptop>#<CWKB92S81IWF.TY7GT6M2EVYC@notmylaptop>
Signed-off-by: Sebastian <sebastian@sebsite.pw>