check: resolve type when taking size/alignment
size/align expressions previously only resolved the dimensions of the
type, which allowed some invalid types to be used without complaint (for
instance, the secondary type of a pointer type was never checked).

This isn't easy to fix, since the size/align expression may be used
within another type, and we want to allow circular dependencies if it
doesn't cause issues when resolving the type's dimensions (for example:
`type t = [size(*t)]int;` should be allowed).

This patch solves the problem by continuing to only resolve dimensions
when the size/alignment expression is first checked, but storing a list
of ast_types which need to be fully resolved later. These types are
resolved in wrap_resolver, and in check_function. The resolution needs
to happen in wrap_resolver (rather than check_internal) so the subunit
scope is preserved, for alias types from imported modules.

Signed-off-by: Sebastian <sebastian@sebsite.pw>