check,gen: use name instead of ident to create global_decl declarations
ident is the custom symbol if it is defined. Using it to create the
declaration causes the symbol to be used as the global variable name in
typedef files. This makes the variable unavailable as its hare name from
other modules.

Additionally, in gen.c gen_global_decl(), the qdefs should use the
declaration symbol if it has one.

Signed-off-by: Alejandro W. Sior <aho@sior.be>