type_store: make builtin_type_error have zero size/alignment
Now that zero-size bindings are allowed, there's not really any reason
for builtin_type_error to have size/alignment 1. The two places this
causes problems are the secondary type of pointers and slices, so logic
needs to be added to prevent spurious errors here. But other than that,
this allows for more lenience for things like offsets in struct fields.

Signed-off-by: Sebastian <sebastian@sebsite.pw>