all: use literal/flexible in place of constant
Also renames the token T_LITERAL to T_NUMBER.

There are two things that haven't been changed yet: the first is the
name of iconst/fconst/rconst, and the second is the name of
tests/00-constants.ha. The latter is because renaming it would require
that everyone rerun ./configure, which isn't a big deal, but would
confuse some people, and since we're gonna be changing harec's build
system when merging it with hare anyways, we may as well just wait for
that so no manual intervention is required at all.

Signed-off-by: Sebastian <sebastian@sebsite.pw>