gen: move tagged value to first aligned offset
This moves the value of a tagged union to its first aligned offset, so
that values of disjointed alignment no longer occupy the same address.
The advantage of this approach is that all tagged unions are
bytewise-compatible with each other regardless of the alignment of their
members, which simplifies the design considerably and opens up
opportunities for some new features.

Layout before:
	(i32 | i64) == struct {
		tag: u32,
		@offset(8) _i32: i32,
		@offset(8) _i64: i64,
	}
	(i32 | u32) == struct {
		tag: u32,
		@offset(4) _i32: i32, // Note offset of this field
		@offset(4) _u32: u32,
	}

Layout after:
	(i32 | i64) == struct {
		tag: u32,
		@offset(4) _i32: i32, // Note offset of this field
		@offset(8) _i64: i64,
	}
	(i32 | u32) == struct {
		tag: u32,
		@offset(4) _i32: i32, // Is independent of alignment
		@offset(4) _u32: u32,
	}