Add missing testmod::s_a allocation
In testmod/testmod.ha, s_x/s_a was defined as a prototype [1]:

    export let @symbol("s_x") s_a: int;

But tests/34-declarations.ha uses it [2]:

    fn imported() void = {
        testmod::s_a;
    }

This currently works on QBE because QBE recognizes the `loadsw $s_x`
return value is never used, and eliminates the expression, but in
fact $s_x is never defined anywhere in the final linked binary
tests_34_declarations.

I found this issue while testing my own QBE implementation, which
does not have this optimization yet.

This patch fixes that.

[1]: https://git.sr.ht/~sircmpwn/harec/tree/770566a51aa972c320b545d2292626057aabe831/item/testmod/testmod.ha#L26
[2]: https://git.sr.ht/~sircmpwn/harec/tree/770566a51aa972c320b545d2292626057aabe831/item/tests/34-declarations.ha#L531