Fix storage union handling in check_expr_literal
Storage is a union of values with different sizes, passing
the contents unconditionally as int64 might work on amd64 but
breaks on other archs.
Since the type information is available we can use the correct
union field for each type instead.

Found the hard way trying to build this on big endian PowerPC.

Signed-off-by: Tobias Heider <me@tobhe.de>