util: ignore write errors to stderr
This is pretty pointless; if writing to stderr fails, we try to, print
an error message to stderr, and then exit. If we're writing to stderr,
an error already occurred, so there's no reason to catch any write
errors here.
~
This is necessary for the upcoming patch, since stderr is closed by
rt::compile but we still want to know where the original error came
from.

Signed-off-by: Sebastian <sebastian@sebsite.pw>