ast: improve representation of array literal expandability
Previously, every element in the literal had an  element 'expand' and
all except the last one was ignored. Now the whole literal holds a single
bool for expandability and the list of array element is storedd in the
common ast_expression_list struct.

References: https://todo.sr.ht/~sircmpwn/hare/889