Implement optional function parameters.
>From Sebastian's RFC. Allows for function prototypes to contain
optional parameters, so for example memio::dynamic and
memio::dynamic_from can become

  fn dynamic(in: []u8 = []) stream =  { ... };

The optional parameters are checked and evaluated at compile time. When
the function is called without optional parameters, they are filled in
at the call site (rather than abusing variadics).

Four features are missing, to be added in future patches.
- Allowing function pointers to be assigned despite disagreeing optional exprs.
- Checking the optional exprs for exported fns don't use non-exported names.
- Non-null optional pointers.
- NaN floats.

Signed-off-by: Joe Finney <me@spxtr.net>