gen: rewrite expandable array -> slice assignment
It only handled a very narrow subset of the intended functionality
previously - assignment from expandable arrays with more than one
element is now possible, as well as assignment of aggregate values.
Slice bounds are properly checked now, and the non-expandable part of
the array is generated in place.

One testcase was removed from tests/08-slices.ha. It was removed
because that operation is not valid at the moment and the new code
catches that correctly - although there are plans to allow this in the
future.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>