check: rewrite check_expr_call
Make variadicity of the last argument a property of the call, not of the
last argument itself. Handle too many/not enough arguments in a more
obvious way. Simplify lowering of empty varargs in case of Hare
variadism and unify variadic argument checking with reqular argument
checking.

Add a check for a new error condition - variadic last argument in
nonvariadic functions, and write a bunch of tests for conditions that we
already check for.

Signed-off-by: Bor Grošelj Simić <bgs@turminal.net>