Allow matching on pointers to tagged unions.
Previously, if you wanted to match on a nullable *(t1 | t2), you
needed to first match out null, then dereference the result and match
again, then work with the resulting values of type t1 and t2. This
patch makes it possible to have cases for null, *t1, and *t2 directly.

Zero-sized types do not match as pointers, so *(int | void) can have
cases *int and void, but not *void.

Signed-off-by: Joe Finney <me@spxtr.net>
Signed-off-by: Drew DeVault <sir@cmpwn.com>