gen: fix ... placement in C-style variadic calls
The ... argument was only added after the first argument (or not added
at all if there were no arguments). This meant that calls to functions
with the signature `(...)` generated incorrect IR. This doesn't cause
issues on x86_64, because the calling convention for variadic and
non-variadic parameters is the same, but it could cause miscompilation
on other platforms.

Likewise, zero size arguments weren't properly handled. In addition to
the ... argument only being added at the end of the loop, param was only
advanced to the next parameter at the end of the loop, so neither would
be reached for arguments with zero size. This caused the ... argument to
either generate in the wrong place or not generate at all if any
non-variadic arguments had zero size.

Both of these issues have been fixed, so a ... argument will always be
generated when appropriate, and always in the correct place.

Signed-off-by: Sebastian <sebastian@sebsite.pw>