lex: s/T_NUMBER/T_LITERAL/g
T_NUMBER was also used for rune and string literals. This could result
in odd error messages, for example, if a non-literal is used as the
argument for @symbol, the error message said it expected a number. This
has been changed to now say it expects a literal.

This still isn't great: for one, "literal" is too vague; ideally the
error message should specify exactly what kind of literal it exepcts.
It's also not entirely accurate: "literal" only refers to number, rune,
and string literals, not to array, struct, tuple, boolean, or unit
(void, done, soon nomem) literals.

Signed-off-by: Sebastian <sebastian@sebsite.pw>